+++ runtime/docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.44] - 2023-06-26
+
+- Add a new ServiceConfig field to servicer config
+
 ## [0.0.0.43] - 2023-06-14
 
 - Rename package import types to coreTypes for consitency and clarity